requirements/common.txt
# Django Core
django==4.2.11
django-extensions==3.2.3
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Web Scraping
requests==2.32.2
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.38.0
webdriver-manager==4.0.2

# Data Analysis
pandas==2.2.2

# Testing
pytest==8.2.0
pytest-django==4.8.0
factory-boy==3.3.0

# Utilities
pillow==10.2.0
python-dateutil==2.8.2